refactor(web): stylesモジュールからnavigation関連を分離
未使用のSTYLES定数とcreateNavigateToHomeActionをstylesから削除。
favorites・mypageのインポート先をnavigation.tsに変更。
